// src/app/models/system-config/setFrontEndData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SetFrontEndData {
    activeFlag?: string;
    certLogoPath?: string;
    contactAddress?: string;
    contactEmail?: string;
    contactTel?: string;
    courseCount1?: number;
    courseCount2?: number;
    courseCount3?: number;
    courseCount4?: number;
    createDate?: Date;
    facebookPath?: string;
    igPath?: string;
    kmitlToken?: string;
    linePath?: string;
    logoFooterPath?: string;
    logoHeaderPath?: string;
    metaDescription?: string;
    moodleToken?: string;
    popularType?: number;
    registerOtpFlag?: boolean;
    registerTokenExpireMinute?: number;
    setId?: number;
    superAdminEmail?: string;
    superAdminName?: string;
    title?: string;
    universityNameEn?: string;
    universityNameTh?: string;
    universityTaxNo?: string;
    updateDate?: Date;
    virtualTranscriptSignPath?: string;
    virtualTranscriptSignPathEn?: string;
    virtualTranscriptSignPathTh?: string;
    website?: string;
    xPath?: string;
    youtubePath?: string;
    popularCourseDisplayLimit?: number;

    logoHeaderPrefix?: string ;
	logoHeaderModule?: number;
	logoFooterPrefix?: string ;
	logoFooterModule?: number;

}